passed-review version

+++ printf/ft_print_int.c
@@ -6,7 +6,7 @@
 /*   By: tkono <tkono@student.42tokyo.jp>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/13 22:27:23 by tkono             #+#    #+#             */
-/*   Updated: 2026/02/21 15:15:12 by tkono            ###   ########.fr       */
+/*   Updated: 2026/02/26 18:58:03 by tkono            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_print_int(int n)
 {
-	int		count;
-	int		tmp;
-	long	num;
+	int			count;
+	int			tmp;
+	long long	num;
 
 	count = 0;
 	num = n;
